Refactor one page for each type of render
- Refactors code to set different pages for list view (continent) and detail view (country).

- Both pages are used by root page to maintain SPA concept.
