+++ src/features/bikeList/BikeList.css
@@ -1,3 +1,3 @@
-.hidden {
-  display: none;
+* {
+  font-size: medium;
 }
